regime-classifier/pyproject.toml
[project]
name = "regime-classifier"
version = "0.1.0"
description = "Market regime classifier for AlgoTrix — 4D feature scoring with Euclidean, HMM, and GMM approaches"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "pandas>=2.1",
    "numpy>=1.26",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "hmmlearn>=0.3",
    "nolds>=0.6",
    "TA-Lib>=0.4.28",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
regime-classifier = "cli:main"
